fixed private cloud export
