+++ queries/paginated/private-cloud/index.ts
@@ -3,7 +3,7 @@ import {
   getPrivateCloudProjectsQuery,
   getPrivateCloudProjectsResult,
   getPrivateCloudProjectsTotalCount,
-} from '@/queries/private-cloud/helpers';
+} from '@/queries/paginated/private-cloud/helpers';
 import { PrivateProject } from '@/queries/types';
 
 export async function privateCloudProjectsPaginated(
@@ -14,7 +14,7 @@ export async function privateCloudProjectsPaginated(
   cluster?: string | null, // Non admins will be required to pass this field that will filter projects for their user
   userEmail?: string | null,
   ministryRoles?: string[],
-  active: boolean = true,
+  active?: boolean | null,
 ): Promise<{
   data: PrivateProject[];
   total: number;
@@ -26,7 +26,7 @@ export async function privateCloudProjectsPaginated(
     cluster,
     userEmail,
     ministryRoles,
-    active,
+    active: !!active,
   });
 
   const proms = [];
