+++ queries/paginated/public-cloud/helpers.ts
@@ -1,23 +1,27 @@
 import prisma from '@/lib/prisma';
-import { PrivateProject } from '@/queries/types';
+import { PublicProject } from '@/queries/types';
 
-export async function getPrivateCloudProjectsQuery({
+export async function getPublicCloudProjectsQuery({
   searchTerm,
   ministry,
-  cluster,
+  provider,
   userEmail,
   ministryRoles,
   active,
 }: {
   searchTerm?: string | null;
   ministry?: string | null;
-  cluster?: string | string[] | null;
+  provider?: string | string[] | null;
   userEmail?: string | null;
   ministryRoles?: string[];
-  active: boolean;
+  active?: boolean;
 }) {
   // Initialize the search/filter query
-  const searchQuery: any = active ? { status: 'ACTIVE' } : {};
+  const searchQuery: any = active
+    ? {
+        status: 'ACTIVE',
+      }
+    : {};
 
   // Construct search/filter conditions based on provided parameters
   if (searchTerm) {
@@ -79,7 +83,7 @@ export async function getPrivateCloudProjectsQuery({
       { name: { $regex: searchTerm, $options: 'i' } },
       { description: { $regex: searchTerm, $options: 'i' } },
       { licencePlate: { $regex: searchTerm, $options: 'i' } },
-      { cluster: { $regex: searchTerm, $options: 'i' } },
+      { provider: { $regex: searchTerm, $options: 'i' } },
       { ministry: { $regex: searchTerm, $options: 'i' } },
 
       // include other fields as necessary
@@ -120,19 +124,19 @@ export async function getPrivateCloudProjectsQuery({
     ];
   }
 
-  if (cluster) {
-    if (Array.isArray(cluster)) {
-      if (cluster.length > 0) searchQuery.cluster = { $in: cluster };
+  if (provider) {
+    if (Array.isArray(provider)) {
+      if (provider.length > 0) searchQuery.provider = { $in: provider };
     } else {
-      searchQuery.cluster = cluster;
+      searchQuery.provider = provider;
     }
   }
 
   return searchQuery;
 }
 
-export async function getPrivateCloudProjectsTotalCount({ searchQuery }: { searchQuery: any }) {
-  const result: unknown = await prisma.privateCloudProject.aggregateRaw({
+export async function getPublicCloudProjectsTotalCount({ searchQuery }: { searchQuery: any }) {
+  const result: unknown = await prisma.publicCloudProject.aggregateRaw({
     pipeline: [
       {
         $lookup: {
@@ -169,7 +173,7 @@ export async function getPrivateCloudProjectsTotalCount({ searchQuery }: { searc
   return 0;
 }
 
-export async function getPrivateCloudProjectsResult({
+export async function getPublicCloudProjectsResult({
   searchQuery,
   skip,
   pageSize,
@@ -182,14 +186,13 @@ export async function getPrivateCloudProjectsResult({
     return [];
   }
 
-  let paginationPipelines: any[] = [];
-  paginationPipelines = [{ $skip: skip }, { $limit: pageSize }];
+  const paginationPipelines = pageSize === 0 ? [] : [{ $skip: skip }, { $limit: pageSize }];
 
-  const result = await prisma.privateCloudProject.aggregateRaw({
+  const result = await prisma.publicCloudProject.aggregateRaw({
     pipeline: [
       {
         $lookup: {
-          from: 'PrivateCloudRequest', // The foreign collection
+          from: 'PublicCloudRequest', // The foreign collection
           let: { projectId: '$_id' }, // Define variable for use in the pipeline
           pipeline: [
             {
@@ -260,5 +263,5 @@ export async function getPrivateCloudProjectsResult({
     ],
   });
 
-  return result as unknown as PrivateProject[];
+  return result as unknown as PublicProject[];
 }
